Add comments in composer, add TODOs, refactor the signatures
